pyproject.toml
[project]
name = "cohere_api"
description = "Examples of calling Cohere API"

requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }

authors = [
    { name = "Miao Ju", email = "jumiao56@gmail.com" },
]

dependencies = [
    "absl-py",
    "cohere",
    "pre-commit",
    "mkdocs",
    "mkdocs-material",
    "ruff",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "B", "I"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.per-file-ignores]
"**/__init__.py" = ["F401"]
